Extract selecting utxo from ogmios results and fix prepare-configuration wizard
